Rename formatDate.js to date.js
I want to add a function `isTomorrow(today)` to the file. Therefore, I
chose a more general name.
